ai-financial-tool/pyproject.toml
[project]
name = "ai-financial-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.4",
    "fastapi>=0.116.1",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "sseclient-py>=1.8.0",
    "openpyxl>=3.1.5",
    "pymongo>=4.14.0",
]

